fix(PAT-94): Move fuction into consultation.
